requirements.txt
dspy-ai>=2.4
pandas>=2.0
tqdm>=4.65
beautifulsoup4>=4.12
python-dateutil>=2.8
google-auth>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.1
google-api-python-client>=2.0
openai>=1.0
pydantic>=2.0
pyarrow==23.0.1
fastparquet==2025.12.0